services/embeddings/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
faiss-cpu==1.7.3
fastapi==0.93.0
filelock==3.9.0
h11==0.14.0
hiredis==2.2.2
httptools==0.5.0
huggingface-hub==0.13.1
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
opensearch-py==2.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
opensearch_dsl